Update Double rating

+++ src/main/java/edu/ncsu/dbms/wolfmedia/services/MetadataService.java
@@ -55,7 +55,7 @@ public class MetadataService {
      * @return true if the query is executed successfully
      * @throws Exception if the query fails
      */
-    public Boolean addPodcastRatings(Integer podcastId, Integer rating) throws Exception {
+    public Boolean addPodcastRatings(Integer podcastId, Double rating) throws Exception {
         String query = "UPDATE podcasts SET rating = " + rating + " WHERE podcastId = " + podcastId;
         return genericDAO.executeUpdate(query);
